Remove AppMode
Allow users to add and delete rooms at any time
