+++ src/main/java/seedu/resireg/MainApp.java
@@ -76,13 +76,10 @@ public class MainApp extends Application {
     private Model initModelManager(Storage storage, ReadOnlyUserPrefs userPrefs) {
         Optional<ReadOnlyResiReg> resiRegOptional;
         ReadOnlyResiReg initialData;
-        boolean isRoomCreationAllowed = true;
         try {
             resiRegOptional = storage.readResiReg();
             if (!resiRegOptional.isPresent()) {
                 logger.info("Data file not found. Will be starting with a sample ResiReg");
-            } else {
-                isRoomCreationAllowed = false; // finalize rooms if succesfully loaded from existing file
             }
             initialData = resiRegOptional.orElseGet(SampleDataUtil::getSampleResiReg);
         } catch (DataConversionException e) {
@@ -96,10 +93,6 @@ public class MainApp extends Application {
         // delete expired items on start up
         ModelManager modelManager = new ModelManager(initialData, userPrefs);
         modelManager.deleteExpiredBinItems();
-        // finalize rooms if necessary
-        if (!isRoomCreationAllowed) {
-            modelManager.finalizeRooms();
-        }
         return modelManager;
     }
