--- src/main/java/seedu/resireg/model/AppMode.java
@@ -1,6 +1,0 @@
-package seedu.resireg.model;
-
-public enum AppMode {
-    NEW,
-    NORMAL;
-}
