+++ src/main/java/seedu/resireg/model/Model.java
@@ -341,8 +341,4 @@ public interface Model {
      * Saves the current ResiReg state for undo/redo.
      */
     void saveStateResiReg();
-
-    AppMode getAppMode();
-
-    void finalizeRooms();
 }
