+++ src/main/resources/view/MainPanel.fxml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import javafx.scene.layout.VBox?>
+<?import java.lang.*?>
+<?import java.util.*?>
+<?import javafx.scene.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
 
-<?import javafx.scene.control.Tab?>
-<?import javafx.scene.layout.StackPane?>
 <?import javafx.geometry.Insets?>
-<?import javafx.scene.control.TabPane?>
-
 <StackPane xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1">
 
     <TabPane fx:id="tabPane"
