+++ src/test/java/seedu/resireg/logic/LogicManagerTest.java
@@ -47,8 +47,6 @@ public class LogicManagerTest {
     @BeforeEach
     public void setUp() {
         model = new ModelManager();
-        model.finalizeRooms();
-
         JsonResiRegStorage resiRegStorage =
                 new JsonResiRegStorage(temporaryFolder.resolve("resireg.json"));
         JsonUserPrefsStorage userPrefsStorage = new JsonUserPrefsStorage(temporaryFolder.resolve("userPrefs.json"));
@@ -92,7 +90,6 @@ public class LogicManagerTest {
                 + FACULTY_DESC_AMY + STUDENT_ID_DESC_AMY;
         Student expectedStudent = new StudentBuilder(AMY).withTags().build();
         ModelManager expectedModel = new ModelManager();
-        expectedModel.finalizeRooms();
         expectedModel.addStudent(expectedStudent);
         expectedModel.saveStateResiReg();
         String expectedMessage = LogicManager.FILE_OPS_ERROR_MESSAGE + DUMMY_IO_EXCEPTION;
