+++ src/test/java/seedu/resireg/logic/commands/AddAliasCommandTest.java
@@ -18,7 +18,6 @@ import javafx.collections.ObservableList;
 import seedu.resireg.commons.core.GuiSettings;
 import seedu.resireg.logic.CommandHistory;
 import seedu.resireg.logic.commands.exceptions.CommandException;
-import seedu.resireg.model.AppMode;
 import seedu.resireg.model.Model;
 import seedu.resireg.model.ModelPredicate;
 import seedu.resireg.model.ReadOnlyResiReg;
@@ -350,16 +349,6 @@ public class AddAliasCommandTest {
 
         @Override
         public void saveStateResiReg() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public AppMode getAppMode() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void finalizeRooms() {
             throw new AssertionError("This method should not be called.");
         }
     }
