+++ src/test/java/seedu/resireg/logic/commands/ClearCommandTest.java
@@ -18,7 +18,6 @@ public class ClearCommandTest {
     @Test
     public void execute_emptyResiReg_success() {
         Model model = new ModelManager();
-        model.finalizeRooms();
         Model expectedModel = new ModelManager();
         expectedModel.saveStateResiReg();
